requirements.txt
beautifulsoup4 
numpy 
psycopg2-binary
